pyproject.toml
[tool.poetry]
name = "docs-manager"
version = "0.1.0"
description = "An End-to-End secured File Managing System."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1.1"
django-encrypted-files = "^0.0.10"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
